docs/D/Demo/A2A-Resource-Processor/resource-replacer/pyproject.toml
[project]
name = "resource-replacer-agent"
version = "0.1.0"
description = "A2A Resource Replacer Agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.6",
    "uvicorn>=0.34.2",
    "httpx>=0.28.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "click>=8.1.8",
    "rich>=13.5.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 